refactor(airflow): extract evidence mappers into _evidence.py
Move map_get_recent_airflow_failures, map_get_airflow_dag_runs, and
map_get_airflow_task_instances out of __init__.py into a dedicated
sibling module, per review feedback that facade files should only
expose the public tool interface.
